requirements.txt
torch
diffusers
transformers
Pillow
numpy
accelerate
sentencepiece
protobuf
boto3
requests
